Added --noS1 flag back to gNB launching.
It was recommended to take it out previously
because it  waas unnecessary. Now that we are
using the EPC as well, we need this flag to
distinguis the mode. When the --noS1 flag is
added, we do not setup the tunnel interface.
Without this flag, the gNB will try to set up
the tunnel interface when NOT in EPC mode and
will crash.