Add --max-seconds option
We now set an alarm to automatically kill this process after some
number of seconds (default: 600 -- 10 minutes).

This alarm is important because we run with the real-time
scheduler.  If (due to bugs) this process were to become run-away
(running continuously without ever blocking), the alarm will
eventually kill the process.  Otherwise, the host machine would
need to be rebooted.

This commit also reworks the command-line argument parsing in
main() to accommodate the new --max-seconds option.