Slowing down PNF to temporarily improve timing
Currently, TX_DATA_REQs seem to be occasionally
taking too long to arrive at the OAI UE from the
proxy (sendto() -> recvfrom()). Also, the slot
indications are not lining up with the control and
data nFAPI indications, which cause ACK/nACK issues,
and ultimately crash the gNB. To temporariliy
mitigate this problem, we show the proxy down for
each slot to better align the nFAPI control/data
inds and the slot indications.

A fix will have to be implemented for these timing
issues soon.