Correcting nfapi parameters passed to test script.
Based on the new NFAPI command line change in
the Eurecom develop branch, commit no. 3f61f27b,
or MR !1868, we need to updated the --nfapi
parameters we are passing into the run-oai
script. Furthermore, we recently removed the
LTE HO changes from the multi-ue-proxy repo.
These changes will be added back into the
proxy when MR !1719 is successfully merged to
the develop branch.